fix: correct sqlc-generated params for MarkEventsAsRecurringByPattern
Add explicit ::float8 casts to $4/$5/$6 in the SQL query so sqlc
correctly infers numeric types. Update the generated params struct
to use float64 (DayOfWeek, TimeOfDay, DurationMinutes) instead of
the incorrect pgtype.Timestamptz. Remove the raw-SQL workaround in
the repository and use the properly-typed generated code instead.

Agent-Logs-Url: https://github.com/waydxd/Orbit-core/sessions/a86b7fb1-c54d-4ded-bf9a-6577c50244e8

+++ sql/queries/habit.sql
@@ -86,7 +86,7 @@ UPDATE events
 SET is_recurring = TRUE, updated_at = $1
 WHERE user_id = $2
   AND LOWER(TRIM(title)) = LOWER(TRIM($3))
-  AND EXTRACT(DOW FROM start_time) = $4
-  AND EXTRACT(HOUR FROM start_time) * 60 + EXTRACT(MINUTE FROM start_time) = $5
-  AND EXTRACT(EPOCH FROM (end_time - start_time))/60 = $6
+  AND EXTRACT(DOW FROM start_time) = $4::float8
+  AND EXTRACT(HOUR FROM start_time) * 60 + EXTRACT(MINUTE FROM start_time) = $5::float8
+  AND EXTRACT(EPOCH FROM (end_time - start_time))/60 = $6::float8
   AND is_recurring = FALSE;
